# docs/swagger.yaml
basePath: /api
definitions:
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.Favorite:
    properties:
      created_at:
        type: string
      id:
        type: integer
      recipe_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.FavoriteRequest:
    properties:
      recipe_id:
        type: integer
      user_id:
        type: integer
    required:
    - recipe_id
    - user_id
    type: object
...
  models.Profile:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      full_name:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
host: religious-mitzi-arasycorp-cc5d6628.koyeb.app
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for culinary review API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API Culinary Review
  version: "1.0"
paths:
...
  /api/favorites:
    post:
      consumes:
      - application/json
      description: Creates a new favorite for a user with the specified recipe ID.
      parameters:
      - description: Favorite data to create
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.FavoriteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Favorite'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Create a new favorite
      tags:
      - favorites
  /api/favorites/{id}:
    delete:
      description: Deletes a favorite by its ID.
      parameters:
      - description: Favorite ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Favorite deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Delete a favorite by ID
      tags:
      - favorites
  /api/favorites/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieves favorites associated with the specified user ID.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Favorite'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Retrieve favorites by user ID
      tags:
      - favorites
...
  /api/profile:
    post:
      consumes:
      - multipart/form-data
      description: Create a new user profile
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Full Name
        in: formData
        name: fullName
        required: true
        type: string
      - description: Bio
        in: formData
        name: bio
        required: true
        type: string
      - description: Avatar File
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Profile'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new profile
      tags:
      - profiles
    put:
      consumes:
      - multipart/form-data
      description: Update the profile of the authenticated user
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Full Name
        in: formData
        name: fullName
        required: true
        type: string
      - description: Bio
        in: formData
        name: bio
        required: true
        type: string
      - description: Avatar File
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update profile by user ID
      tags:
      - profiles
  /api/profile/me:
    get:
      consumes:
      - application/json
      description: Get the profile of the authenticated user
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Profile'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get profile by user ID
      tags:
      - profiles
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"